+++ packages/node-core/src/index.ts
@@ -3,7 +3,6 @@
 
 export * from './api.service';
 export * from './logger';
-export * from './yargs';
 export * from './profiler';
 export * from './events';
 export * from './configure';