+++ packages/node/src/utils/project.ts
@@ -24,10 +24,12 @@ import {
   SubstrateHandler,
   SubstrateHandlerKind,
 } from '@subql/common-substrate';
+import { StoreService } from '@subql/node-core';
+import { getAllEntitiesRelations } from '@subql/utils';
 import yaml from 'js-yaml';
 import tar from 'tar';
 import { NodeVM, VMScript } from 'vm2';
-import { SubqlProjectDs } from '../configure/SubqueryProject';
+import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
 
 export async function prepareProjectDir(projectPath: string): Promise<string> {
   const stats = fs.statSync(projectPath);
@@ -298,4 +300,13 @@ export function loadChainTypesFromJs(
     );
   }
   return rawContent;
+}
+
+export async function initDbSchema(
+  project: SubqueryProject,
+  schema: string,
+  storeService: StoreService,
+): Promise<void> {
+  const modelsRelation = getAllEntitiesRelations(project.schema);
+  await storeService.init(modelsRelation, schema);
 }